feat: :construction: Se aplico la validacion de delete en los modulos existentes

+++ js/modules/posts.js
@@ -7,7 +7,7 @@ export const getAllPosts = async() => {
     return data;
 }
 
-const validar = (arg) => {
+const validarAdd = (arg) => {
     if (Object.keys(arg).length !== 3) {
         console.error("El argumento NO cumple la validacion: 'Numero de Argumentos'");
         return false;
@@ -23,9 +23,16 @@ const validar = (arg) => {
     return true;
 }
 
+const validarDelete = (id) => {
+    if (typeof id !== "string") {
+        return false;
+    }
+    return true;
+}
+
 export const addPost = async(arg) => {
 
-    if (validar(arg)) {
+    if (validarAdd(arg)) {
         let userId = arg.userId;
         let title = arg.title;
         let body = arg.body;
@@ -49,11 +56,14 @@ export const addPost = async(arg) => {
 }
 
 export const deletePost = async(id)=>{
-    let config = {
-    method: "DELETE",
-    headers: {"Content-Type": "application/json"}
+    if (validarDelete(id)) {
+        let config = {
+            method: "DELETE",
+            headers: {"Content-Type": "application/json"}
+            }
+            let res = await fetch(`${remoto.posts}/${id}`, config);
+            let data = await res.json();
+            return data;
     }
-    let res = await fetch(`${remoto.posts}/${id}`, config);
-    let data = await res.json();
-    return data;
+    return false;
 }
